backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
motor>=3.3.0
pymongo>=4.5.0
python-dotenv>=1.0.0
dnspython>=2.4.0